+++ packages/web-threejs/CHANGELOG.md
@@ -1,3 +1,14 @@
+## [Unreleased]
+
+### Changed
+
+- Updated README.md documentation for assets installation process
+- Updated LICENSE.md license information
+
+### Fixed
+
+- Fixed Ready Player Me bug with Three.js r160
+
 ## [1.0.0] - 2023-12-12
 
 - Initial Release